requirements.txt
tensorflow==2.10.0
keras==2.10.0
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
scikit-image==0.21.0
Flask==3.0.3
flask_cors==5.0.0
Pillow==10.4.0
gunicorn==23.0.0
flask_wtf==1.2.1
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.24
flask_login==0.6.3
email_validator==2.2.0
b2sdk>=2,<3
python-dotenv==1.0.1
pytest==8.3.4